Separate garages state query as a partial class
