+++ NineChronicles.Headless/GraphTypes/StateQuery.cs
@@ -26,7 +26,7 @@ using NineChronicles.Headless.GraphTypes.States.Models.Table;
 
 namespace NineChronicles.Headless.GraphTypes
 {
-    public class StateQuery : ObjectGraphType<StateContext>
+    public partial class StateQuery : ObjectGraphType<StateContext>
     {
         public StateQuery()
         {
@@ -542,31 +542,7 @@ namespace NineChronicles.Headless.GraphTypes
                 }
             );
 
-            Field<GarageStateType>(
-                "garage",
-                arguments: new QueryArguments(
-                    new QueryArgument<NonNullGraphType<AddressType>>
-                    {
-                        Name = "address",
-                        Description = "Address to get GARAGE token balance and fungible items"
-                    },
-                    new QueryArgument<ListGraphType<StringGraphType>>
-                    {
-                        Name = "fungibleItemIds",
-                        Description = "List of fungible item IDs to get stock in garage"
-                    }
-                ),
-                resolve: context =>
-                {
-                    var address = context.GetArgument<Address>("address");
-                    var balance = context.Source.GetBalance(address, Currencies.Garage);
-                    var fungibleItemIdList = context.GetArgument<IEnumerable<string>>("fungibleItemIds");
-                    IEnumerable<Address> fungibleItemAddressList = fungibleItemIdList.Select(fungibleItemId =>
-                        Addresses.GetGarageAddress(address, HashDigest<SHA256>.FromString(fungibleItemId)));
-                    var fungibleItemList = context.Source.GetStates(fungibleItemAddressList.ToArray());
-                    return (balance, fungibleItemList);
-                }
-            );
+            RegisterGarages();
         }
     }
 }
